Starting models refactor
